requirements.txt
flask[async]
openai
pyyaml